+++ shared/profile/link-with-icon.js.flow
@@ -1,6 +1,6 @@
 // @flow
 import React, {Component} from 'react'
-import type {IconType} from './icon.constants.js'
+import type {IconType} from '../common-adapters/icon.constants.js'
 
 export type Props = {
   label: string,
