+++ shared/profile/post-proof.desktop.js
@@ -1,7 +1,8 @@
 // @flow
 import * as shared from './post-proof.shared'
 import * as React from 'react'
-import {Box, Button, CopyableText, Icon, LinkWithIcon, PlatformIcon, Text} from '../common-adapters'
+import {Box, Button, CopyableText, Icon, PlatformIcon, Text} from '../common-adapters'
+import LinkWithIcon from './link-with-icon'
 import {clipboard} from 'electron'
 import {globalStyles, globalColors, globalMargins, desktopStyles} from '../styles'
